fix: change AdminPolygon vertex_count from uint16 to uint32
The uint16_t vertex_count field truncated polygons at 65,535 vertices,
corrupting boundaries like Deutschland (217K verts) and Bayern (113K)
when simplification is disabled or uses very small epsilon.

Struct stays 24 bytes (the uint16 had 2 bytes padding after it that
the uint32 now fills naturally). No change to disk/memory footprint.

Updated in: types.h, s2_helpers.cpp, build_index.cpp,
continent_filter.cpp, simplify_compare.cpp, server main.rs

+++ builder/src/continent_filter.cpp
@@ -111,7 +111,7 @@ ParsedData filter_by_bbox(const ParsedData& full, const ContinentBBox& bbox) {
           AdminPolygon nap = ap;
           nap.vertex_offset = static_cast<uint32_t>(out.admin_vertices.size());
           out.admin_polygons.push_back(nap);
-          for (uint16_t v = 0; v < ap.vertex_count; v++)
+          for (uint32_t v = 0; v < ap.vertex_count; v++)
               out.admin_vertices.push_back(full.admin_vertices[ap.vertex_offset + v]);
       } }
